+++ rando/static/js/popup.js
@@ -18,10 +18,16 @@ $(document).ready(function (e) {
         $modal.modal('show');
     }
 
-    // Click on home shows popup
-    $('header a.home.popup').click(function () {
-        $modal.modal('show');
+    $(window).on('view:home view:detail', function () {
+        // Popup is shown only on homepage
+        $('header a.home.popup, header li.home a').click(function (e) {
+            if (/^\/[a-zA-Z_]{2,5}\/$/.test(window.location.pathname)) {
+                $modal.modal('show');
+                e.preventDefault();
+            }
+        });
     });
+
 
     function showModal() {
         var alreadyShown = localStorage.getItem('popup-shown') === "yes",